+++ Packages/ai.natml.vision.movenet-3d/Runtime/MoveNet3DPose.cs
@@ -1,6 +1,6 @@
 /*
 *   MoveNet 3D
-*   Copyright (c) 2022 NatML Inc. All Rights Reserved.
+*   Copyright © 2023 NatML Inc. All Rights Reserved.
 */
 
 namespace NatML.Vision {