+++ Packages/ai.natml.vision.movenet-3d/Changelog.md
@@ -1,2 +1,7 @@
+## 1.0.1
++ Added `MoveNet3DPredictor.Create` static method for creating the predictor.
++ Removed `MoveNetPredictor` public constructor.
++ Upgraded to NatML 1.1.2.
+
 ## 1.0.0
 + First release.